@@ -12,22 +12,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.9",
-    "@angular/cdk": "17.3.9",
-    "@angular/common": "17.3.9",
-    "@angular/compiler": "17.3.9",
-    "@angular/core": "17.3.9",
-    "@angular/forms": "17.3.9",
-    "@angular/material": "17.3.9",
-    "@angular/platform-browser": "17.3.9",
-    "@angular/platform-browser-dynamic": "17.3.9",
-    "@angular/router": "17.3.9",
+    "@angular/animations": "18.2.4",
+    "@angular/cdk": "18.2.3",
+    "@angular/common": "18.2.4",
+    "@angular/compiler": "18.2.4",
+    "@angular/core": "18.2.4",
+    "@angular/forms": "18.2.4",
+    "@angular/material": "18.2.3",
+    "@angular/platform-browser": "18.2.4",
+    "@angular/platform-browser-dynamic": "18.2.4",
+    "@angular/router": "18.2.4",
     "@fontsource/roboto": "^4.5.8",
-    "@ng-doc/app": "17.6.13",
-    "@ng-doc/builder": "17.6.13",
-    "@ng-doc/core": "17.6.13",
-    "@ng-doc/ui-kit": "17.6.13",
-    "@nx/angular": "18.2.4",
+    "@ng-doc/app": "18.0.3",
+    "@ng-doc/builder": "18.0.3",
+    "@ng-doc/core": "18.0.3",
+    "@ng-doc/ui-kit": "18.0.3",
+    "@nx/angular": "19.7.2",
     "backoff-rxjs": "^7.0.0",
     "chart.js": "^4.3.3",
     "chartjs-adapter-date-fns": "^3.0.0",
@@ -42,17 +42,19 @@
     "rxjs": "~7.5.0",
     "tinycolor2": "^1.5.2",
     "tslib": "^2.6.2",
-    "zone.js": "0.14.7"
+    "zone.js": "0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.8",
-    "@angular-devkit/core": "17.3.8",
-    "@angular-devkit/schematics": "17.3.8",
-    "@angular/cli": "~17.3.0",
-    "@angular/compiler-cli": "17.3.9",
-    "@nx/eslint": "^19.2.3",
-    "@nx/workspace": "18.2.4",
-    "@schematics/angular": "17.3.8",
+    "@angular-devkit/build-angular": "18.2.4",
+    "@angular-devkit/core": "18.2.4",
+    "@angular-devkit/schematics": "18.2.4",
+    "@angular/cli": "~18.2.0",
+    "@angular/compiler-cli": "18.2.4",
+    "@nx/eslint": "19.7.2",
+    "@nx/workspace": "19.7.2",
+    "@schematics/angular": "18.2.4",
+    "@types/d3": "^7.4.3",
+    "@types/dompurify": "^3.0.5",
     "@types/fhir": "^0.0.37",
     "@types/jasmine": "~4.0.0",
     "@types/lodash-es": "^4.17.6",
@@ -67,13 +69,13 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
     "lint-staged": "^14.0.1",
-    "ng-packagr": "17.3.0",
-    "nx": "18.2.4",
+    "ng-packagr": "18.2.1",
+    "nx": "19.7.2",
     "postcss": "^8.4.28",
     "postcss-import": "~14.1.0",
     "postcss-preset-env": "~7.5.0",
     "postcss-url": "~10.1.3",
     "prettier": "2.7.1",
-    "typescript": "5.4.5"
+    "typescript": "5.5.4"
   }
 }
